web/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
gevent==23.9.1
gevent-websocket==0.10.1
flask-socketio==5.3.6
eventlet==0.33.3
paho-mqtt==1.6.1
python-socketio==5.11.1